对组织机构模块进行重构

+++ src/Services/Organization/Surging.Hero.Organization.Domain/Organizations/Departments/DepartmentDomainService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Surging.Core.AutoMapper;
@@ -23,16 +24,19 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
         private readonly IDapperRepository<Department, long> _departmentRepository;
         private readonly IDapperRepository<Corporation, long> _corporationRepository;
         private readonly IDapperRepository<Position, long> _positionRepository;
+        private readonly IDapperRepository<Organization, long> _organizationRepository;
         private readonly IPositionDomainService _positionDomainService;
     
         public DepartmentDomainService(IDapperRepository<Department, long> departmentRepository,
             IDapperRepository<Corporation, long> corporationRepository,
             IDapperRepository<Position, long> positionRepository,
+            IDapperRepository<Organization, long> organizationRepository,
             IPositionDomainService positionDomainService)
         {
             _departmentRepository = departmentRepository;
             _corporationRepository = corporationRepository;
             _positionRepository = positionRepository;
+            _organizationRepository = organizationRepository;
             _positionDomainService = positionDomainService;
            
         }
@@ -56,27 +60,35 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
                 throw new BusinessException($"部门类型Id不正确,请选择正确的部门类型");
             }
             var department = input.MapTo<Department>();
+            var orgInfo = input.MapTo<Organization>();
             var thisDepartmentCount = await _departmentRepository.GetCountAsync(p => p.CorporationId == input.CorporationId);
-
-            if (input.ParentId == 0)
+            var parentOrg = await _organizationRepository.SingleOrDefaultAsync(p => p.Id == input.ParentId);
+            if (parentOrg == null)
             {
-                department.Level = 1;
-                department.Code = corporation.Code + HeroConstants.CodeRuleRestrain.CodeSeparator + (thisDepartmentCount + 1).ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
+                throw new BusinessException($"不存在Id为{input.ParentId}的上级信息");
             }
             else
             {
-                var parentDept = await _departmentRepository.SingleOrDefaultAsync(p => p.Id == input.ParentId);
-                if (parentDept == null)
+                var orgCode = string.Empty;
+                var maxLevelOrg = (await _organizationRepository.GetAllAsync(p => p.ParentId == parentOrg.Id && p.OrgType == Shared.Organizations.OrganizationType.Corporation)).OrderByDescending(p => p.Id).FirstOrDefault();
+                if (maxLevelOrg == null)
                 {
-                    throw new BusinessException($"系统中不存在Id为{input.ParentId}的父级部门信息");
+                    orgCode = "1".PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
                 }
-                department.Level = parentDept.Level + 1;
-                department.Code = parentDept.Code + HeroConstants.CodeRuleRestrain.CodeSeparator + (thisDepartmentCount + 1).ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
+                else
+                {
+                    orgCode = (Convert.ToInt32(maxLevelOrg.Code.TrimStart('0')) + 1).ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
+                }
+
+                orgInfo.Level = parentOrg.Level + 1;
+                orgInfo.Code = parentOrg.Code + HeroConstants.CodeRuleRestrain.CodeSeparator + (thisDepartmentCount + 1).ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
 
             }
 
             await UnitOfWorkAsync(async (conn, trans) =>
             {
+                var orgId = await _organizationRepository.InsertAndGetIdAsync(orgInfo, conn, trans);
+                department.OrgId = orgId;
                 var deptId = await _departmentRepository.InsertAndGetIdAsync(department, conn, trans);
                 if (input.Postions != null && input.Postions.Any())
                 {
@@ -88,7 +100,7 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
                     {
                         position.DeptId = deptId;
                         position.CheckDataAnnotations().CheckValidResult();
-                        var positionCode = department.Code + HeroConstants.CodeRuleRestrain.CodeSeparator + sort.ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
+                        var positionCode = orgInfo.Code + HeroConstants.CodeRuleRestrain.CodeSeparator + sort.ToString().PadLeft(HeroConstants.CodeRuleRestrain.CodeCoverBit, HeroConstants.CodeRuleRestrain.CodeCoverSymbol);
                         await _positionDomainService.CreatePosition(position, positionCode, conn, trans);
                         sort++;
                     }
@@ -96,24 +108,34 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
             }, Connection);
         }
 
-        public async Task DeleteDepartment(long id)
+        public async Task DeleteDepartmentByOrgId(long orgId)
         {
-            var department = await _departmentRepository.SingleOrDefaultAsync(p => p.Id == id);
-            if (department == null)
+            var department = await _departmentRepository.SingleOrDefaultAsync(p => p.OrgId == orgId);
+            var orgInfo = await _organizationRepository.SingleOrDefaultAsync(p => p.Id == orgId);
+            if (department == null || orgInfo == null)
             {
-                throw new BusinessException($"系统中不存在Id为{id}的部门信息");
+                throw new BusinessException($"系统中不存在Id为{orgId}的部门信息");
             }
-            var children = await _departmentRepository.GetAllAsync(p => p.ParentId == id);
+            var children = await _organizationRepository.GetAllAsync(p => p.ParentId == orgId);
             if (children.Any()) {
                 throw new BusinessException($"请先删除子部门信息");
             }
             var departmentUsers = await GetService<IUserAppService>().GetDepartmentUser(department.Id);
-            if (departmentUsers.Any()) {
-                throw new BusinessException($"该部门存在用户,请先删除该部门下的用户");
-            }
+            //if (departmentUsers.Any()) {
+            //    throw new BusinessException($"该部门存在用户,请先删除该部门下的用户");
+            //}
             await UnitOfWorkAsync(async (conn, trans) => {
+                await _organizationRepository.DeleteAsync(orgInfo, conn, trans);
                 await _departmentRepository.DeleteAsync(department,conn,trans);
-                await _positionRepository.DeleteAsync(p => p.Code.Contains(department.Code), conn, trans);
+                await _positionRepository.DeleteAsync(p => p.Code.Contains(orgInfo.Code), conn, trans);
+                foreach (var departmentUser in departmentUsers)
+                {
+                    if (!await GetService<IUserAppService>().ResetUserOrgInfo(departmentUser.Id))
+                    {
+                        throw new BusinessException("重置该公司部门员工部门信息失败,请稍后重试");
+                    }
+                }
+
             },Connection);
             
         }
@@ -121,11 +143,37 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
         public async Task<GetDepartmentOutput> GetDepartment(long id)
         {
             var department = await _departmentRepository.SingleOrDefaultAsync(p => p.Id == id);
+           
             if (department == null)
             {
                 throw new BusinessException($"系统中不存在Id为{id}的部门信息");
             }
+            var orgInfo = await _organizationRepository.SingleOrDefaultAsync(p => p.Id == department.OrgId);
+            if (orgInfo == null) {
+                throw new BusinessException($"系统中不存在Id为{id}的部门信息");
+            }
             var departmentOutput = department.MapTo<GetDepartmentOutput>();
+            departmentOutput = orgInfo.MapTo(departmentOutput);
+            var positions = await _positionDomainService.GetPositionsByDeptId(department.Id);
+            departmentOutput.Postions = positions.MapTo<IEnumerable<GetPositionOutput>>();
+            return departmentOutput;
+        }
+
+        public async Task<GetDepartmentOutput> GetDepartmentByOrgId(long orgId)
+        {
+            var department = await _departmentRepository.SingleOrDefaultAsync(p => p.OrgId == orgId);
+
+            if (department == null)
+            {
+                throw new BusinessException($"系统中不存在orgId为{orgId}的部门信息");
+            }
+            var orgInfo = await _organizationRepository.SingleOrDefaultAsync(p => p.Id == department.OrgId);
+            if (orgInfo == null)
+            {
+                throw new BusinessException($"系统中不存在orgId为{orgId}的部门信息");
+            }
+            var departmentOutput = department.MapTo<GetDepartmentOutput>();
+            departmentOutput = orgInfo.MapTo(departmentOutput);
             var positions = await _positionDomainService.GetPositionsByDeptId(department.Id);
             departmentOutput.Postions = positions.MapTo<IEnumerable<GetPositionOutput>>();
             return departmentOutput;
@@ -134,13 +182,20 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
         public async Task UpdateDepartment(UpdateDepartmentInput input)
         {
             var department = await _departmentRepository.SingleOrDefaultAsync(p => p.Id == input.Id);
+           
             if (department == null) {
                 throw new BusinessException($"系统中不存在Id为{input.Id}的部门信息");
             }
+            var orgInfo = await _organizationRepository.SingleOrDefaultAsync(p => p.Id == department.OrgId);
+            if (orgInfo == null) {
+                throw new BusinessException($"系统中不存在Id为{department.Id}的部门信息");
+            }
+
             if (input.DeptTypeId == 0)
             {
                 throw new BusinessException($"请选择部门类型");
             }
+          
             var workbookAppServiceProxy = GetService<IWordbookAppService>();
             var checkDeptTypeResult = await workbookAppServiceProxy.Check(new CheckWordbookInput() { WordbookCode = SystemPresetWordbookCode.Organization.DeptType, WordbookItemId = input.DeptTypeId });
             if (!checkDeptTypeResult)
@@ -148,7 +203,12 @@ namespace Surging.Hero.Organization.Domain.Organizations.Departments
                 throw new BusinessException($"部门类型Id不正确,请选择正确的部门类型");
             }            
             department = input.MapTo(department);
-            await _departmentRepository.UpdateAsync(department);
+            orgInfo = input.MapTo(orgInfo);
+
+            await UnitOfWorkAsync(async (conn, trans) => {
+                await _organizationRepository.UpdateAsync(orgInfo, conn, trans);
+                await _departmentRepository.UpdateAsync(department, conn, trans);
+            }, Connection);
         }
     }
 }
